refactor: major site improvements and reorganization
- Move content from src/data/blog to content/ for better separation
- Replace custom middleware with Astro's built-in redirects config
- Reorganize utils/ into focused subdirectories (posts/, images/, text/, markdown/)
- Remove Alpine.js references from documentation (now using vanilla JS)
- Add MIT LICENSE with dual licensing (content vs source code)
- Remove "Recent Posts" section from homepage
- Add emojis to About page headers and Source Code section
- Update all import paths for reorganized utilities
