+++ astro.config.ts
@@ -8,12 +8,15 @@ import {
   transformerNotationHighlight,
   transformerNotationWordHighlight,
 } from "@shikijs/transformers";
-import { transformerFileName } from "./src/utils/transformers/fileName.js";
+import { transformerFileName } from "./src/utils/markdown/transformers/fileName.js";
 import { SITE } from "./src/config";
 
 // https://astro.build/config
 export default defineConfig({
   site: SITE.website,
+  redirects: {
+    "/blog/[...slug]": "/posts/[...slug]"
+  },
   integrations: [
     sitemap({
       filter: page => SITE.showArchives || !page.endsWith("/archives"),
