+++ content/blog/assets/thinking-big-and-small/clear.png
