+++ content/blog/assets/why-i-didnt-apply-to-ycombinator/columbia-admissions-email.png
