+++ content/blog/entrepreneur-interviews-dan-shipper.md
