+++ content/blog/entrepreneur-interviews-greg-skloot.md
