+++ content/blog/entrepreneur-interviews-jack-mcdermott.md
