+++ content/blog/going-all-in.md
