+++ content/blog/in-defense-of-knowing-less.md
