+++ content/blog/six-months-as-a-dropout.md
