+++ content/blog/sucking-less-at-business-development.md
