+++ content/blog/the-5-am-principle.md
