+++ content/blog/the-future-is-service-oriented.md
