+++ content/blog/the-open-default-principle.md
