+++ content/blog/thinking-big-and-small.md
