+++ content/blog/why-i-didnt-apply-to-ycombinator.md
