+++ src/pages/about.md
@@ -3,18 +3,22 @@ layout: ../layouts/AboutLayout.astro
 title: "About"
 ---
 
-## Code
+## 💻 Code
 
 You can find my programming projects on [GitHub](https://github.com/bendrucker). I work on a variety of open source projects, mainly in Go and JavaScript.
 
-## Cycling
+## 🚴 Cycling
 
 Find my rides on [Strava](https://www.strava.com/athletes/5723594). I am a longtime member and current president of the [San Francisco Cycling Club](https://www.sfcyclingclub.org).
 
-## Work
+## 💼 Work
 
 Founding engineer at _Stealth_.
 
-## Contact
+## 📧 Contact
 
 The best way to reach me is via [email](mailto:bvdrucker@gmail.com). I'm not actively posting to any of my social profiles.
+
+## 🔧 Source Code
+
+The source code for this website is available on GitHub ([`bendrucker/bendrucker.me`](https://github.com/bendrucker/bendrucker.me)), under the [MIT license](https://github.com/bendrucker/bendrucker.me/blob/master/LICENSE). Pull requests are welcome if you notice a typo!
