+++ src/pages/archives/index.astro
@@ -5,7 +5,7 @@ import Layout from "@/layouts/Layout.astro";
 import Header from "@/components/Header.astro";
 import Footer from "@/components/Footer.astro";
 import Card from "@/components/Card.astro";
-import getPostsByGroupCondition from "@/utils/getPostsByGroupCondition";
+import getPostsByGroupCondition from "@/utils/posts/group";
 import { SITE } from "@/config";
 
 // Redirect to 404 page if `showArchives` config is false
