+++ src/pages/og.png.ts
@@ -1,5 +1,5 @@
 import type { APIRoute } from "astro";
-import { generateOgImageForSite } from "@/utils/generateOgImages";
+import { generateOgImageForSite } from "@/utils/images/generate";
 
 export const GET: APIRoute = async () =>
   new Response(await generateOgImageForSite(), {
