+++ src/pages/posts/[...page].astro
@@ -7,7 +7,7 @@ import Header from "@/components/Header.astro";
 import Footer from "@/components/Footer.astro";
 import Card from "@/components/Card.astro";
 import Pagination from "@/components/Pagination.astro";
-import getSortedPosts from "@/utils/getSortedPosts";
+import getSortedPosts from "@/utils/posts/sort";
 import { SITE } from "@/config";
 
 export const getStaticPaths = (async ({ paginate }) => {
