+++ src/pages/posts/[...slug]/index.astro
@@ -1,8 +1,8 @@
 ---
 import { type CollectionEntry, getCollection } from "astro:content";
 import PostDetails from "@/layouts/PostDetails.astro";
-import getSortedPosts from "@/utils/getSortedPosts";
-import { getPath } from "@/utils/getPath";
+import getSortedPosts from "@/utils/posts/sort";
+import { getPath } from "@/utils/posts/path";
 
 export interface Props {
   post: CollectionEntry<"blog">;
