+++ src/pages/tags/index.astro
@@ -5,7 +5,7 @@ import Layout from "@/layouts/Layout.astro";
 import Tag from "@/components/Tag.astro";
 import Header from "@/components/Header.astro";
 import Footer from "@/components/Footer.astro";
-import getUniqueTags from "@/utils/getUniqueTags";
+import { getUniqueTags } from "@/utils/posts/tags";
 import { SITE } from "@/config";
 
 const posts = await getCollection("blog");
