+++ src/utils/posts/sort.ts
@@ -1,5 +1,5 @@
 import type { CollectionEntry } from "astro:content";
-import postFilter from "./postFilter";
+import postFilter from "./filter";
 
 const getSortedPosts = (posts: CollectionEntry<"blog">[]) => {
   return posts
